+++ src/plugins/relay/api/remote/relay-remote-event.h
@@ -20,6 +20,8 @@
 #ifndef WEECHAT_PLUGIN_RELAY_REMOTE_EVENT_H
 #define WEECHAT_PLUGIN_RELAY_REMOTE_EVENT_H
 
+#include <cjson/cJSON.h>
+
 #define RELAY_REMOTE_EVENT_ID_INITIAL_SYNC "initial_sync"
 
 #define RELAY_REMOTE_EVENT_CALLBACK(__body_type)                        \
@@ -43,6 +45,8 @@ struct t_relay_remote_event_cb
     t_relay_remote_event_func *func;    /* callback (can be NULL)           */
 };
 
+extern void relay_remote_event_buffer_input (struct t_gui_buffer *buffer,
+                                             const char *input_data);
 extern void relay_remote_event_recv (struct t_relay_remote *remote,
                                      const char *data);