+++ CHANGELOG.md
@@ -73,6 +73,7 @@
 - relay/api: fix "body_type" returned when lines or nicks of a buffer are requested
 - relay/api: fix read of one buffer line
 - relay/api: automatically disconnect when the command `/upgrade` is executed on the remote
+- relay/api: fix send of data to remote after command `/upgrade` executed in the remote client
 - relay/api: disconnect from remote in case of error when sending data
 - relay/api: disconnect cleanly when the remote is quitting ([#2168](https://github.com/weechat/weechat/issues/2168))
 - relay: fix websocket permessage-deflate extension when the client doesn't send the max window bits parameters ([#1549](https://github.com/weechat/weechat/issues/1549))